+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.5.0] - 2020-06-09
+### Added
+* ObservableArray now has the ability to "toggle" an item (remove if present, else add)
+### Fixed
+* Observables will now notify even when the value is set to something "roughly" equivalent (for example, when an observable value is changed from `true` to `1`)
+
 ## [2.4.0] - 2020-05-12
 ### Added
 * Added the ability to directly remove an item from an ObservableArray